feat: Send CloudEvents for Watch Events
Updating the "Job" controller to send a CloudEvent when an object is
created, updated, or deleted. This requires the controller to be
configured with a receiver to send events to. CloudEvents are chosen
because they are a standard way to emit "events" in cloud-native
applications. These events are agnostic to the receiver and can be
integrated into other frameworks, like KNative Eventing.

The current implementation is a proof of concept that emits very simple
data. Standardizing the format/API of the transmitted event data will
come later. The proof of concept also includes a test-receiver binary,
which acts as a dummy receiver of the emitted CloudEvents.

Signed-off-by: Adam Kaplan <adam@adambkaplan.com>