Dynamic Controllers from Config
Rename controller to "dynamic" and utilize k8s "Unstructured" objects
to set up the controller and read data from the cluster. This allows
any Kubernetes object to be reconciled with the same fundamental
code. Controller-runtime only cares about the GroupVersionKind when
setting up the clients, informers, etc. for the controller. A "Name"
field was also added to facilitate logging.

For configuration, spf13/viper and spf13/pflag were added to load
settings from config files, flags, and environment variables. This
is most useful for the "watched" objects, which can be defined in a
YAML file. The "main" function was enhanced to bind flags to the
Viper configuration, and load configuration from a file if present.
Viper also handles precedence if a configuration is defined more than
once - see https://github.com/spf13/viper for more information.

Documentation was also updated to include a reference for the config
YAML file, and a "try it" appeal to developers.

Signed-off-by: Adam Kaplan <adam@adambkaplan.com>